+++ src/utils/callkeep.ts
@@ -190,7 +190,6 @@ export const setupCallKeepEvents = async () => {
     console.log('CallKeep debug: didDeactivateAudioSession')
     BrekekeUtils.webrtcSetAudioEnabled(false)
     cs.updateBackgroundCalls()
-
   }
   const didReceiveStartCallAction = async (
     e: EventsPayload['didReceiveStartCallAction'],
@@ -212,7 +211,7 @@ export const setupCallKeepEvents = async () => {
       await waitSip()
       getCallStore().startCall(e.handle)
     }
- }
+  }
 
   // https://github.com/react-native-webrtc/react-native-callkeep#didloadwithevents
   const didLoadWithEventsHandlers: { [k: string]: Function } = {